renamed llmproxy folder and filename to proxyllm
