+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "llmproxy"
+name = "proxyllm"
 version = "0.1.0"
 description = "LLM Proxy to reduce cost and complexity of using multiple LLMs"
 authors = ["laxitshahi"]
